refactor: extraer lista de vista y coordinar componentes en vista de trails
